moved getSHA1 util fun from app module to shared module
