+++ shared/src/iosMain/kotlin/org/listenbrainz/shared/util/PlatformUtils.ios.kt
@@ -1,0 +1,9 @@
+package org.listenbrainz.shared.util
+
+import org.listenbrainz.shared.repository.PlatformContext
+
+actual object PlatformUtils{
+    actual fun getSHA1(context: PlatformContext,packageName:String):String?{
+        return null
+    }
+}
